Alternative extra typings
