+++ types/postcss-cssnext.d.ts
@@ -1,0 +1,14 @@
+import postcss = require('postcss');
+
+interface Options {
+	console?: typeof console;
+	warnForDuplicates?: boolean;
+	warnForDuplication?: boolean;
+	features?: {[name: string]: any};
+	browsers: string[];
+}
+
+// TODO: fix Options
+declare function cssnext(options: any): postcss.Plugin<Options>;
+
+export = cssnext;
